Fixing the broken.. Tiny steps at a time ;-)
* Fixing major issues (major issues that might have been there from
early days)
* fixed the workflow up to IPN notifying the server and generating
success msg to the buyer (then: crashes)
* Some refactoring
* Typos
* Version bump
