+++ spree_paypal_website_standard.gemspec
@@ -1,8 +1,10 @@
 # encoding: UTF-8
+version = File.read(File.expand_path("../GEM_VERSION",__FILE__)).strip
+
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'spree_paypal_website_standard'
-  s.version     = '1.0.0.alfa1'
+  s.version     = version
   s.summary     = 'Spree extension for integration with PayPal Website Standard payment'
   s.description = 'Spree extension for integration with PayPal Website Standard payment'
   s.required_ruby_version = '>= 1.9.2'
